Timer, progressbar, tab margin, window resize

+++ qt5/box.cpp
@@ -26,7 +26,7 @@ void uiBoxAppend(uiBox *b, uiControl *c, int stretchy)
 void uiBoxDelete(uiBox *b, int index)
 {
 	if (auto layoutBox = uiValidateAndCastObjTo<QBoxLayout>(b)) {
-		if (index < (uint)layoutBox->count()) {
+		if (index < layoutBox->count()) {
 			delete layoutBox->takeAt(index);
 		}
 	}
